requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
cloudflare==2.1.0
Flask==1.0.2
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsonlines==1.2.0
MarkupSafe==1.0
python-dotenv==0.8.2
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
